Remove DEP, add description for remove_padding, and move around tests
